moved sass files to assets/sass
