+++ src/assets/sass/components/Header/IconDelete.module.scss
