+++ src/assets/sass/components/Workspace/_preview.module.scss
