+++ src/assets/sass/components/_workspace.module.scss
