+++ src/components/Header/DeleteButton.tsx
@@ -6,7 +6,7 @@ import {
     deleteActiveDocument,
     selectActiveDocumentId,
 } from '@/store/documentsSlice';
-import styles from '@/styles/components/Header/IconDelete.module.scss';
+import styles from '@/assets/sass/components/Header/IconDelete.module.scss';
 
 function DeleteButton() {
     // todo: documentMap[id] can be undefined for current doc id
