+++ src/components/Workspace/Splitter.tsx
@@ -1,4 +1,4 @@
-import styles from '@/styles/components/_workspace.module.scss';
+import styles from '@/assets/sass/components/_workspace.module.scss';
 
 function Splitter() {
     return <div className={`w-px min-h-full ${styles.splitter}`}></div>;
